Fixed using All Config not changing vanilla items to the base item model
